requirements.txt
keyboard
pyserial
setuptools==65.3.0
pycaw
pyyaml
watchdog